refactor(gui): 起動ブランチ resolver を共有化
